[MCUX-72474] Fixed compiler error in distanceFunctions
[Status]
Error[Pa093]: implicit conversion from floating point to integer.
CMSIS\DSP\Source\DistanceFunctions\arm_dtw_path_f32.c - Line 152,156.

+++ DSP/Source/DistanceFunctions/arm_dtw_path_f32.c
@@ -149,11 +149,11 @@ ARM_DSP_ATTRIBUTE void arm_dtw_path_f32(const arm_matrix_instance_f32 *pDTW,
   {
      itemp = fh[0];
      fh[0] = sh[0];
-     sh[0] = itemp;
+     sh[0] = (int16_t)itemp;
 
      itemp = fh[1];
      fh[1] = sh[1];
-     sh[1] = itemp;
+     sh[1] = (int16_t)itemp;
 
      fh += 2;
      sh -= 2;
